requirements.txt
numpy
librosa
google-genai
pydantic
python-dotenv
torch
torchaudio
encodec
matplotlib
plotly
pandas
